votes: Move voting logic to sg_votes.cpp
