+++ src/sgame/sg_main.cpp
@@ -1997,198 +1997,12 @@ void CheckExitRules()
 }
 
 /*
-==================
-G_Vote
-==================
-*/
-void G_Vote( gentity_t *ent, team_t team, bool voting )
-{
-	if ( !level.team[ team ].voteTime )
-	{
-		return;
-	}
-
-	//TODO maybe refactor vote no/yes in one and only one variable and divide the NLOC by 2 ?
-	if ( voting )
-	{
-		if ( (ent->client->pers.voted & ( 1 << team )) )
-		{
-			return;
-		}
-		level.team[ team ].voted++;
-		ent->client->pers.voted |= 1 << team;
-		if ( ent->client->pers.voteYes & ( 1 << team ) )
-		{
-			level.team[ team ].voteYes++;
-			trap_SetConfigstring( CS_VOTE_YES + team, va( "%d", level.team[ team ].voteYes ) );
-		}
-
-		if ( ent->client->pers.voteNo & ( 1 << team ) )
-		{
-			level.team[ team ].voteNo++;
-			trap_SetConfigstring( CS_VOTE_NO + team, va( "%d", level.team[ team ].voteNo ) );
-		}
-		return;
-	}
-
-	if ( !( ent->client->pers.voted & ( 1 << team ) ) )
-	{
-		return;
-	}
-
-	ent->client->pers.voted &= ~( 1 << team );
-	level.team[ team ].voted--;
-	bool wasYes = ent->client->pers.voteYes & ( 1 << team );
-	bool wasNo  = ent->client->pers.voteNo  & ( 1 << team );
-	ent->client->pers.voteYes &= ~( 1 << team );
-	ent->client->pers.voteNo  &= ~( 1 << team );
-
-	if ( wasYes )
-	{
-		level.team[ team ].voteYes--;
-		trap_SetConfigstring( CS_VOTE_YES + team, va( "%d", level.team[ team ].voteYes ) );
-	}
-	if ( wasNo )
-	{
-		level.team[ team ].voteNo--;
-		trap_SetConfigstring( CS_VOTE_NO + team, va( "%d", level.team[ team ].voteNo ) );
-	}
-}
-
-void G_ResetVote( team_t team )
-{
-	int i;
-
-	level.team[ team ].voteTime = 0;
-	level.team[ team ].voteYes = 0;
-	level.team[ team ].voteNo = 0;
-	level.team[ team ].voted = 0;
-
-	for ( i = 0; i < level.maxclients; i++ )
-	{
-		level.clients[ i ].pers.voted &= ~( 1 << team );
-		level.clients[ i ].pers.voteYes &= ~( 1 << team );
-		level.clients[ i ].pers.voteNo &= ~( 1 << team );
-	}
-
-	trap_SetConfigstring( CS_VOTE_TIME + team, "" );
-	trap_SetConfigstring( CS_VOTE_STRING + team, "" );
-	trap_SetConfigstring( CS_VOTE_YES + team, "0" );
-	trap_SetConfigstring( CS_VOTE_NO + team, "0" );
-}
-/*
 ========================================================================
 
 FUNCTIONS CALLED EVERY FRAME
 
 ========================================================================
 */
-
-void G_ExecuteVote( team_t team )
-{
-	level.team[ team ].voteExecuteTime = 0;
-
-	trap_SendConsoleCommand( va( "%s", level.team[ team ].voteString ) );
-
-	if ( !Q_stricmp( level.team[ team ].voteString, "map_restart" ) )
-	{
-		G_MapLog_Result( 'r' );
-		level.restarted = true;
-	}
-	else if ( !Q_strnicmp( level.team[ team ].voteString, "map", 3 ) )
-	{
-		G_MapLog_Result( 'm' );
-		level.restarted = true;
-	}
-}
-
-
-/*
-==================
-G_CheckVote
-==================
-*/
-void G_CheckVote( team_t team )
-{
-	float    votePassThreshold = ( float ) level.team[ team ].voteThreshold / 100.0f;
-	bool pass = false;
-	bool quorum = true;
-	char     *cmd;
-
-	if ( level.team[ team ].voteExecuteTime /* > 0 ?? more readable imho */ &&
-	     level.team[ team ].voteExecuteTime < level.time )
-	{
-		G_ExecuteVote( team );
-	}
-
-	if ( !level.team[ team ].voteTime )
-	{
-		return;
-	}
-
-	if ( ( level.time - level.team[ team ].voteTime >= VOTE_TIME ) ||
-	     ( level.team[ team ].voteYes + level.team[ team ].voteNo == level.team[ team ].numPlayers ) )
-	{
-		pass = ( level.team[ team ].voteYes &&
-		         ( float ) level.team[ team ].voteYes / ( ( float ) level.team[ team ].voteYes + ( float ) level.team[ team ].voteNo ) > votePassThreshold );
-	}
-	else
-	{
-		if ( ( float ) level.team[ team ].voteYes >
-		     ( float ) level.team[ team ].numPlayers * votePassThreshold )
-		{
-			pass = true;
-		}
-		else if ( ( float ) level.team[ team ].voteNo <=
-		          ( float ) level.team[ team ].numPlayers * ( 1.0f - votePassThreshold ) )
-		{
-			return;
-		}
-	}
-
-	// If quorum is required, check whether at least half of who could vote did
-	if ( level.team[ team ].quorum && level.team[ team ].voted < floor( powf( level.team[ team ].numPlayers, 0.6 ) ) )
-	{
-		quorum = false;
-	}
-
-	if ( pass && quorum )
-	{
-		level.team[ team ].voteExecuteTime = level.time + level.team[ team ].voteDelay;
-	}
-
-	G_LogPrintf( "EndVote: %s %s %d %d %d %d",
-	             team == TEAM_NONE ? "global" : BG_TeamName( team ),
-	             pass ? "pass" : "fail",
-	             level.team[ team ].voteYes, level.team[ team ].voteNo, level.team[ team ].numPlayers, level.team[ team ].voted );
-
-	if ( !quorum )
-	{
-		cmd = va( "print_tr %s %d %d", ( team == TEAM_NONE ) ? QQ( N_("Vote failed ($1$ of $2$; quorum not reached)") ) : QQ( N_("Team vote failed ($1$ of $2$; quorum not reached)") ),
-		            level.team[ team ].voteYes + level.team[ team ].voteNo, level.team[ team ].numPlayers );
-	}
-	else if ( pass )
-	{
-		cmd = va( "print_tr %s %d %d", ( team == TEAM_NONE ) ? QQ( N_("Vote passed ($1$ — $2$)") ) : QQ( N_("Team vote passed ($1$ — $2$)") ),
-		            level.team[ team ].voteYes, level.team[ team ].voteNo );
-	}
-	else
-	{
-		cmd = va( "print_tr %s %d %d %.0f", ( team == TEAM_NONE ) ? QQ( N_("Vote failed ($1$ — $2$; $3$% needed)") ) : QQ( N_("Team vote failed ($1$ — $2$; $3$% needed)") ),
-		            level.team[ team ].voteYes, level.team[ team ].voteNo, votePassThreshold * 100 );
-	}
-
-	if ( team == TEAM_NONE )
-	{
-		trap_SendServerCommand( -1, cmd );
-	}
-	else
-	{
-		G_TeamCommand( team, cmd );
-	}
-
-	G_ResetVote( team );
-}
 
 static void SetNeedpass(const std::string& password)
 {
